SAK-29180 Remove the PDAHandler, refactor, and cleanup. (1) Remove and unregister PDA handler, refactor into SiteHandler and clean up SkinnableCharonPortal (2) Further removal of PDA code and rename of the sakai.properties that control inlining (3) Cleanup the property naming - remove obsolete properties and update documentation. (4) Remove some obsolete constants (5) Update hierarchy documentation to remove incorrect stuff
Removed:
portal.pda.autoredirect=false
portal.tablets.use.mobile=false
portal.frameset=

portal.frameset was obsolete long ago - I just took it out of the default properties
and updated the doc while I was in the files.
