+++ config/configuration/bundles/src/bundle/org/sakaiproject/config/bundle/default.sakai.properties
@@ -572,20 +572,19 @@
 # DEFAULT: 
 # portal.neoprefix=
 
-# If set to false, auto redirect for mobile devices is disabled and classic view will be used preferentially.
-# TODO: SAK-29115 remove this completely in the code when neo is removed
-# DEFAULT: false
-# portal.pda.autoredirect=false
-
-# SAK-21334 - If set to true (and the portal.pda.autoredirect is also true) will also redirect larger 'tablet' devices to the mobile portal
-# TODO: SAK-29115 remove this completely in the code when neo is removed
-# DEFAULT: false
-# portal.tablets.use.mobile=false
-
-# To override the default regular expression add this property.  
-# To specify a different regular expression for a particular Sakai tool, use a property with this pattern: portal.pda.bypass.[toolid]=
+# This property controls when to bypasss the portal inlining process based on the URL.
+# Often this is the file suffix - but other patterns can be used.
+# This is a regular expression with vertical bars to indicate "or"
 # DEFAULT: \\.jpg$|\\.gif$|\\.js$|\\.png$|\\.jpeg$|\\.prf$|\\.css$|\\.zip$|\\.pdf\\.mov$|\\.json$|\\.jsonp$\\.xml$|\\.ajax$|\\.xls$|\\.xlsx$|\\.doc$|\\.docx$|uvbview$|linktracker$|hideshowcolumns$
-# portal.pda.bypass= 
+# portal.bypass= 
+# To specify a different regular expression than the overall default for a particular Sakai tool, 
+# use a property with this pattern: 
+# portal.bypass.[toolid]=
+
+# This property controls when to bypasss the portal inlining process based on the query parameters
+# This is a regular expression with vertical bars to indicate "or"
+# DEFAULT: wicket:interface=.*iresourcelistener:|wicket:ajax=true
+# portal.bypass.query=
 
 # When true - it runs Tidy in noisy mode. It does *not* take the output of Tidy, just checks the portlet output and then uses it unchanged. 
 # DEFAULT: false
@@ -687,18 +686,17 @@
 # DEFAULT: false
 # portal.experimental.auto.reset=true
 
-# Controls the use of iframes in the portal. If the string contains the text ":all:", then by default all iframes will be suppressed, 
-# and only the tools whose ids are specified will be displayed with iframes.
-# DEFAULT: none (null)
-# portal.experimental.iframesuppress=:all:sakai.assignment:
+# Controls the use of iframes in the portal. It is a colon-separated list
+# of common tool ids like sakai.assignment
+# If the string contains the text ":all:", then by default all iframes 
+# will be suppressed, and only the tools whose ids are listed will 
+# be displayed with iframes.
+# DEFAULT: :all:sakai.rsf.evaluation
+# portal.iframesuppress=
 
 # Whether or not to maximize the display when only one site is available.
 # DEFAULT: false
 # portal.experimental.maximizesinglepage=true
-
-# Configuration option indicaing support for frame set mode. Values (always|never)
-# DEFAULT: none (null) = never
-# portal.frameset=
 
 # Prefix of the handler to use.
 # DEFAULT: site
@@ -800,7 +798,8 @@
 # DEFAULT: true
 # portal.use.profile=false
 
-# Allows for adding addditional code into the header of both the standard and pdaa portals, for example for adding in kaltura or other javascript code
+# Allows for adding additional code into the header of both the standard and pda portals, 
+# for example for adding in kaltura or other javascript code
 # DEFAULT: Empty, no value
 # portal.include.extrahead=<script type="text/javascript" language="JavaScript" src="/kaltura/javascript/kaltura-display.js"></script>
